BundleProvisionerCustomizer classes moved to own files
- Also some minor improvements
- Cleaner filtering files (just syntax with side effects replaced)
- Removed using reflection to call constructors of known classes
